# resources/order-api/openapi-v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Qammaris Order API (Website ↔ Qammaris App)",
        "version": "1.0.0-rc.4.2",
        "description": "r4.2 baseline matching docs/integrations/QAMMARIS_ORDER_API_V1.md r4.2, signed off by the App agent (draft ff3532c).\nr4.2 (Owner decisions 2026-10-09): Issue.opened_by nullable + opened_by_source; waiver\ndecision_id with Owner authorization vs relayed resends (X-Qammaris-Delivery); error codes aligned\n(structure -> validation_failed, approval requirement -> proof_required, unauthorized waiver -> 403);\nOrderListPage.unavailable and serialization_failed so one unreadable order cannot fail the list.\nr4.1 history: K-A (webhook signature path + vectors) and K-B (new timestamp and signature on every webhook retry).\nThe Website is the source of truth for orders; Qammaris App for employees, expenses and reimbursements.\nThe App backend verifies the employee is active and is `owner` or holds `orders.handle` before signing; that\npermission is not sent. The Website webhook (retried up to 24 h) is the primary path; the App reconciles\nopportunistically with `updated_since` because it has no cron.\ntests/Unit/OrderApiContractTest.php validates $refs, the Markdown examples, enums and per-operation error codes.\n"
    },
    "servers": [
        {
            "url": "https://{website}/integrations/qammaris-app/orders/v1",
            "variables": {
                "website": {
                    "default": "example.invalid"
...
    "security": [
        {
            "qammarisSignature": []
        }
    ],
...
    "paths": {
...
        "/orders/{id}/claims/{task}": {
            "delete": {
                "operationId": "releaseClaim",
                "summary": "Holder releases own claim; releasing another holder's claim is owner-only with a reason",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderId"
                    },
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    },
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Task"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReleaseClaimRequest"
                            }
                        }
                    }
                },
                "x-error-codes": [
                    "order_not_found",
                    "action_not_allowed",
                    "revision_conflict",
                    "invalid_transition"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Mutation"
                    },
                    "400": {
                        "$ref": "#/components/responses/Error"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "413": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error"
                    },
                    "default": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/orders/{id}/jnt": {
            "post": {
                "operationId": "recordJnt",
                "summary": "picked_up also sets handover=jnt in the same revision and event; tracking number never required",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderId"
                    },
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JntRequest"
                            }
                        }
                    }
                },
                "x-error-codes": [
                    "order_not_found",
                    "action_not_allowed",
                    "revision_conflict",
                    "invalid_transition"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Mutation"
                    },
                    "400": {
                        "$ref": "#/components/responses/Error"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "413": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error"
                    },
                    "default": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/orders/{id}/issues": {
            "post": {
                "operationId": "openIssue",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderId"
                    },
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IssueRequest"
                            }
                        }
                    }
                },
                "x-error-codes": [
                    "order_not_found",
                    "action_not_allowed",
                    "revision_conflict",
                    "invalid_transition"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Mutation"
                    },
                    "400": {
                        "$ref": "#/components/responses/Error"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "413": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error"
                    },
                    "default": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
        "/orders/{id}/issues/{issue_id}/resolve": {
            "post": {
                "operationId": "resolveIssue",
                "summary": "Resolving an issue opened by someone else is owner-only",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderId"
                    },
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    },
                    {
                        "name": "issue_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ulid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResolveIssueRequest"
                            }
                        }
                    }
                },
                "x-error-codes": [
                    "order_not_found",
                    "action_not_allowed",
                    "revision_conflict",
                    "invalid_transition"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Mutation"
                    },
                    "400": {
                        "$ref": "#/components/responses/Error"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error"
                    },
                    "403": {
                        "$ref": "#/components/responses/Error"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "413": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error"
                    },
                    "default": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "qammarisSignature": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Qammaris-Signature",
                "description": "HMAC-SHA256 over timestamp, method, path+query and body hash, with X-Qammaris-Client and X-Qammaris-Timestamp; see contract §3."
            }
        },
        "parameters": {
            "OrderId": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "$ref": "#/components/schemas/Ulid"
                }
            },
            "IdempotencyKey": {
                "name": "Idempotency-Key",
                "in": "header",
                "required": true,
                "schema": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{16,128}$"
                }
            },
...
        },
        "responses": {
            "Mutation": {
                "description": "Applied, no-op, or idempotent replay",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/MutationResponse"
                        }
                    }
                }
            },
...
        },
        "schemas": {
            "Ulid": {
                "type": "string",
                "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
            },
...
            "Task": {
                "type": "string",
                "enum": [
                    "preparation",
                    "courier_booking",
                    "handover"
                ]
            },
...
            "ReleaseClaimRequest": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "actor"
                ],
                "properties": {
                    "reason": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 200,
                        "description": "Required when an owner releases another holder's claim"
                    },
                    "expected_revision": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "actor": {
                        "$ref": "#/components/schemas/Actor"
                    }
                }
            },
...
        }
    }
}